ebook_management/requirements.txt
Django==3.2
djangorestframework==3.11.0
requests
jsonpath

